fix(router): skip *.client.* files in page-route scanning (#971 P1)
A file like pages/search-widget.client.tsx was both bundled as a
client-script entry AND accepted by the router's page scan as a
/search-widget.client route, because .tsx/.jsx are accepted page
extensions. Client scripts export no page default, so the build/render
could fail or ship an unintended route.

scan_pages now skips any path matching the .client.* filename contract.
To keep that contract a single source of truth without forcing zfb-router
to depend on the heavier zfb-islands crate, the infix/extensions and the
is_client_script_file / client_script_entry_name predicates move into a
new zfb_types::client_scripts module (the zero-zfb-dep leaf both crates
already depend on). zfb-islands re-consumes them via re-export so its
public API is unchanged, and zfb-build's policy.rs is_client_script_path
now delegates to the same predicate instead of duplicating the literals.
