+++ crates/zfb-types/src/lib.rs
@@ -5,6 +5,7 @@
 
 pub mod asset_urls;
 pub mod base_prefix;
+pub mod client_scripts;
 pub mod helpers;
 pub mod segment;
 
@@ -15,5 +16,8 @@ pub use asset_urls::{
     STABLE_ISLANDS_URL,
 };
 pub use base_prefix::dev_mount_prefix;
+pub use client_scripts::{
+    client_script_entry_name, is_client_script_file, CLIENT_SCRIPT_EXTENSIONS, CLIENT_SCRIPT_INFIX,
+};
 pub use helpers::{escape_html, json_string, normalize_path_lexical, path_to_posix_string};
 pub use segment::Segment;
